pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "deep-hook"
version = "3.0.0"
description = "AI-powered code review library — accepts GitLab-style diffs, returns structured review results"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "anshdeep"}
]
keywords = ["code-review", "ai", "gitlab", "langgraph", "langchain"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Quality Assurance",
]
dependencies = [
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",
    "langchain-ollama>=0.2.0",
]

[project.optional-dependencies]
openai = ["langchain-openai>=0.2.0"]
anthropic = ["langchain-anthropic>=0.2.0"]
all = [
    "langchain-openai>=0.2.0",
    "langchain-anthropic>=0.2.0",
]

[project.urls]
Homepage = "https://github.com/anshd258/deep-hook"
Repository = "https://github.com/anshd258/deep-hook"
Issues = "https://github.com/anshd258/deep-hook/issues"

[tool.setuptools.packages.find]
include = ["deep_hook*"]
